# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SEKOR-BKC API
  description: |
    REST API for SEKOR-BKC (শেকড় - The Kolkata Chronicle)
    A Bengali cultural heritage platform for Kolkata's stories.
  version: 1.0.0
  contact:
    name: SEKOR-BKC Team
    email: api@sekor-bkc.com
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.sekor-bkc.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/profile:
    get:
      tags: [auth]
      operationId: getProfile
      summary: Get user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
        '401':
          description: Unauthorized